garmin_app/requirements.txt
garminconnect>=0.2.10
garth>=0.4.40
google-genai>=0.1.0
requests>=2.31.0